# sample_components/co_swagger_0.0.7.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.7",
    "title": "Customer Onboarding At RAK Bank",
    "termsOfService": "http://rakbank.ae/api/terms/",
    "contact": {
      "email": "api-support@rakbank.ae"
    },
    "description": "A typical customer onboarding process involves multiple activities. It starts with evaluating the customer's requirement, followed by due-diligence, completion of legal aspects, account opening and making it operational."
  },
  "host": "rakbank.ae",
  "basePath": "/api/customer-onboarding/v1",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/banks/RAK/usertypes/{userType}/datalist": {
      "get": {
        "summary": "Get list of values with group Ids",
        "operationId": "getDatalist",
        "tags": [
          "Get Datalist"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userType",
            "in": "path",
            "required": true,
            "type": "string",
            "enum": [
              "sme",
              "retail",
              "corporate"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/DatalistResponse"
            }
          },
          "400": {
            "description": "Invalid userType"
          },
          "500": {
            "description": "500 is the generic REST API error response.",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DatalistResponse": {
      "type": "object",
      "required": [
        "GroupId"
      ],
      "properties": {
        "groupId": {
          "type": "array",
          "items": {
            "type": "object",
            "required": [
              "key",
              "value",
              "description"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "key": {
                "type": "string",
                "description": "save value"
              },
              "value": {
                "type": "string"
              },
              "displayText": {
                "type": "string"
              },
              "subGroup": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": [
                    "key",
                    "value",
                    "description"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string",
                      "description": "save value"
                    },
                    "value": {
                      "type": "string"
                    },
                    "displayText": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "errorType": {
          "type": "string",
          "enum": [
            "FieldsValidation",
            "OTP",
            "Other"
          ]
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "fieldPath": {
                "type": "string"
              },
              "errorCode": {
                "type": "string"
              },
              "errorType": {
                "type": "string",
                "enum": [
                  "mandatory",
                  "invalid"
                ]
              },
              "message": {
                "type": "string"
              },
              "developerText": {
                "type": "string",
                "description": "pattern, maxlength, min, max etc used for invalid fields "
              }
            }
          }
        }
      }
    }
  }
}